bugsFixed

+++ Yara/Models/ViewModels/PracticesItem.cs
@@ -41,8 +41,8 @@ namespace Yara.Models.ViewModels
         public string StartTime { set; get; }
         public string FinishDate { set; get; }
         public string FinishTime { set; get; }
-        public int Score { set; get; }
-        public int ScoreBase { set; get; }
+        public double Score { set; get; }
+        public double ScoreBase { set; get; }
         public bool IsFileAnswer { set; get; }
         public string RegDate { set; get; }
         public string RegTime { set; get; }
